ReconstructionByCamera/StyTR-2/requirements.txt
pillow
numpy
scipy
tqdm
tensorboardX
matplotlib